pyproject.toml
[project]
name = "agent-improvements-framework"
version = "1.0.0"
description = "Production-ready AI agent framework with 7 critical improvements for Google ADK agents"
authors = [
    {name = "Kurt", email = "kurt@mapache.ai"},
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10,<3.13"
keywords = ["ai", "agents", "google-adk", "production", "observability", "reliability"]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core Google ADK and Cloud Platform
    "google-adk>=1.15.0,<2.0.0",
    "google-cloud-aiplatform[evaluation,agent-engines]>=1.118.0,<2.0.0",
    "google-cloud-logging>=3.12.0,<4.0.0",
    "google-cloud-trace>=1.11.0,<2.0.0",

    # Observability
    "structlog>=24.1.0,<25.0.0",
    "opentelemetry-api>=1.20.0,<2.0.0",
    "opentelemetry-sdk>=1.20.0,<2.0.0",
    "opentelemetry-exporter-gcp-trace>=1.9.0,<2.0.0",
    "prometheus-client>=0.19.0,<1.0.0",

    # Memory & Data
    "numpy>=1.24.0,<2.0.0",
    "pyyaml>=6.0.0,<7.0.0",

    # Optional Redis support (fakeredis for testing)
    "fakeredis>=2.20.0,<3.0.0",

    # Utilities
    "protobuf>=4.21.0,<5.0.0",
]

[project.optional-dependencies]
# Development dependencies
dev = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<1.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "pytest-timeout>=2.2.0,<3.0.0",
    "nest-asyncio>=1.6.0,<2.0.0",
    "coverage>=7.4.0,<8.0.0",
]

# Code quality
lint = [
    "ruff>=0.4.0,<1.0.0",
    "mypy>=1.8.0,<2.0.0",
    "black>=24.0.0,<25.0.0",
    "isort>=5.13.0,<6.0.0",
    "codespell>=2.2.0,<3.0.0",
    "types-pyyaml>=6.0.12,<7.0.0",
    "types-requests>=2.31.0,<3.0.0",
]

# Production dependencies
prod = [
    "redis>=5.0.0,<6.0.0",  # Real Redis for production
    "psutil>=5.9.0,<6.0.0",  # System monitoring
]

# Jupyter notebooks
jupyter = [
    "jupyter>=1.0.0,<2.0.0",
    "ipykernel>=6.29.0,<7.0.0",
    "ipywidgets>=8.1.0,<9.0.0",
]

# All optional dependencies
all = [
    "redis>=5.0.0,<6.0.0",
    "psutil>=5.9.0,<6.0.0",
    "pytest>=8.0.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<1.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "ruff>=0.4.0,<1.0.0",
    "mypy>=1.8.0,<2.0.0",
    "black>=24.0.0,<25.0.0",
]

[project.urls]
Homepage = "https://github.com/mapachekurt/mapachev1"
Documentation = "https://github.com/mapachekurt/mapachev1/blob/main/README.md"
Repository = "https://github.com/mapachekurt/mapachev1"
Issues = "https://github.com/mapachekurt/mapachev1/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# ===== Pytest Configuration =====
[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=85",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

# ===== Coverage Configuration =====
[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/examples/*",
]
branch = true

[tool.coverage.report]
precision = 2
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

# ===== Ruff Configuration =====
[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests"]
extend-exclude = [
    ".venv",
    "venv",
    "build",
    "dist",
    "*.egg-info",
]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "W",    # pycodestyle warnings
    "I",    # isort
    "C",    # flake8-comprehensions
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
    "RUF",  # ruff specific rules
    "N",    # pep8-naming
    "YTT",  # flake8-2020
    "ASYNC",# flake8-async
    "S",    # flake8-bandit
    "BLE",  # flake8-blind-except
    "A",    # flake8-builtins
    "COM",  # flake8-commas
    "DTZ",  # flake8-datetimez
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "C901",   # too complex
    "B006",   # mutable default argument
    "S101",   # use of assert
    "S311",   # pseudo-random generators
    "COM812", # trailing comma (conflicts with formatter)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",  # asserts allowed in tests
    "S106",  # hardcoded passwords in tests
]
"examples/**/*.py" = [
    "S101",  # asserts allowed in examples
    "T201",  # print allowed in examples
]

[tool.ruff.lint.isort]
known-first-party = ["src"]
force-single-line = false
lines-after-imports = 2

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

# ===== Black Configuration =====
[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# ===== isort Configuration =====
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
known_first_party = ["src"]

# ===== MyPy Configuration =====
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
check_untyped_defs = true
disallow_untyped_defs = false  # Set to true for stricter checking
disallow_incomplete_defs = true
disallow_untyped_calls = false  # Set to true for stricter checking
no_implicit_optional = true
strict_optional = true
follow_imports = "normal"
ignore_missing_imports = true
explicit_package_bases = true
namespace_packages = true
show_error_codes = true
show_column_numbers = true
pretty = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# ===== Codespell Configuration =====
[tool.codespell]
skip = "*.lock,.venv,venv,build,dist,*.egg-info,uv.lock"
ignore-words-list = "rouge,nd"
count = true
quiet-level = 3
